requirements.txt
asgiref
Django
django-environ
django-filter
djangorestframework
pillow
Pygments
pytz
sqlparse
tzdata